UseCase 1:1 매칭
